:hammer: Refactor Ktor plugin and client code for multi-platform support
