+++ composeApp/src/commonMain/kotlin/com/crosspaste/net/plugin/SignalConfig.kt
